requirements.txt
fastapi
uvicorn
python-multipart
deepface
spotipy
nltk
pydantic
Pillow
aiofiles
requests
transformers
torch
